Convert temperatures to Ints

+++ vapi4k-core/src/main/kotlin/com/vapi4k/dsl/model/AnthropicModelImpl.kt
@@ -51,7 +51,7 @@ interface AnthropicModelProperties {
   /**
   This is the temperature that will be used for calls. Default is 0 to leverage caching for lower latency.
    */
-  var temperature: Int
+  var temperature: Double
 
   /**
   These are the tools that the assistant can use during the call. To use transient tools, use `tools`.
