+++ vapi4k-core/src/main/kotlin/com/vapi4k/dsl/model/OpenAIModelImpl.kt
@@ -64,7 +64,7 @@ interface OpenAIModelProperties {
   /**
   This is the temperature that will be used for calls. Default is 0 to leverage caching for lower latency.
    */
-  var temperature: Int
+  var temperature: Double
 
   /**
   These are the tools that the assistant can use during the call. To use transient tools, use `tools`.
