+++ vapi4k-core/src/main/kotlin/com/vapi4k/dtos/model/PerplexityAIModelDto.kt
@@ -26,7 +26,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PerplexityAIModelDto(
-  override var temperature: Int = -1,
+  override var temperature: Double = -1.0,
   override var maxTokens: Int = -1,
   override var emotionRecognitionEnabled: Boolean? = null,
   override var numFastTurns: Int = -1,
